python-service/requirements.txt
Flask==3.0.0
openpyxl==3.1.2
pdfplumber==0.10.3
pandas==2.1.3
numpy==1.26.0
opencv-python==4.8.1.78
tabula-py==2.9.0
